+++ src/renderer/widgets/file-explorer/icons/index.ts
@@ -3,8 +3,12 @@
  * GNU General Public License v3.0 or later (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
  */
 
+import collapseAllSvg from './collapse-all.svg';
+import refreshSvg from './refresh.svg';
 import widgetSvg from './widget.svg';
 
 export {
+  collapseAllSvg,
+  refreshSvg,
   widgetSvg
 }